Feature/backend/battle

See merge request s08-final/S08P31A605!32

+++ backend/ninucco/src/main/java/co/ninuc/ninucco/api/dto/response/MemberRes.java
@@ -13,16 +13,16 @@ public class MemberRes {
 
     String url;
 
-    Long winCount;
+    int winCount;
 
-    Long loseCount;
+    int loseCount;
 
-    Long point;
+    long point;
 
-    Long rate;
+    double rate;
 
     @Builder
-    public MemberRes(String id, String nickname,String url, Long winCount, Long loseCount, Long point, Long rate){
+    public MemberRes(String id, String nickname,String url, int winCount, int loseCount, long point, double rate){
         this.id=id;
         this.nickname=nickname;
         this.url=url;
